requirements.txt
python-gitlab>=2.3.1
python-dateutil>=2.8.0
pyyaml>=5.3.1